improve organization of stylus-core
