+++ stylus-core/src/calls/context.rs
@@ -1,22 +1,12 @@
 // Copyright 2022-2024, Offchain Labs, Inc.
 // For licensing, see https://github.com/OffchainLabs/stylus-sdk-rs/blob/main/licenses/COPYRIGHT.md
 
-use crate::calls::{CallContext, MutatingCallContext, NonPayableCallContext, StaticCallContext};
+use crate::{
+    calls::{CallContext, MutatingCallContext, NonPayableCallContext, StaticCallContext},
+    storage::TopLevelStorage,
+};
 use alloy_primitives::U256;
 use cfg_if::cfg_if;
-
-/// Trait for top-level storage types, usually implemented by proc macros.
-/// Top-level types are special in that their lifetimes track the entirety
-/// of all the EVM state-changes throughout a contract invocation.
-///
-/// To prevent storage aliasing during reentrancy, you must hold a reference
-/// to such a type when making an EVM call. This may change in the future
-/// for programs that prevent reentrancy.
-///
-/// # Safety
-///
-/// The type must be top-level to prevent storage aliasing.
-pub unsafe trait TopLevelStorage {}
 
 /// Enables configurable calls to other contracts.
 #[derive(Debug, Clone)]
