+++ stylus-core/src/calls/mod.rs
@@ -3,6 +3,9 @@
 
 use alloy_primitives::{Address, U256};
 
+pub mod context;
+pub mod errors;
+
 /// Trait for accessing a safe API for calling other contracts.
 /// Its implementation should have reentrancy awareness and protections depending
 /// on the SDK configuration.
@@ -13,7 +16,7 @@ pub trait CallAccess {
         context: &dyn StaticCallContext,
         to: Address,
         data: &[u8],
-    ) -> Result<Vec<u8>, Error>;
+    ) -> Result<Vec<u8>, errors::Error>;
     /// Delegate calls the contract at the given address.
     ///
     /// # Safety
@@ -26,14 +29,14 @@ pub trait CallAccess {
         context: &dyn MutatingCallContext,
         to: Address,
         data: &[u8],
-    ) -> Result<Vec<u8>, Error>;
+    ) -> Result<Vec<u8>, errors::Error>;
     /// Calls the contract at the given address.
     fn call(
         &self,
         context: &dyn MutatingCallContext,
         to: Address,
         data: &[u8],
-    ) -> Result<Vec<u8>, Error>;
+    ) -> Result<Vec<u8>, errors::Error>;
 }
 
 /// Trait for transferring ETH.
@@ -65,7 +68,7 @@ pub trait StaticCallContext: CallContext {}
 ///
 /// The type must contain a [`TopLevelStorage`][TLS] to prevent aliasing in cases of reentrancy.
 ///
-/// [TLS]: stylus_core::context::TopLevelStorage
+/// [TLS]: stylus_core::storage::TopLevelStorage
 pub unsafe trait MutatingCallContext: CallContext {
     /// Amount of ETH in wei to give the other contract.
     fn value(&self) -> U256;
@@ -76,54 +79,3 @@ pub unsafe trait MutatingCallContext: CallContext {
 ///
 /// Note: any implementations of this must return zero for [`MutatingCallContext::value`].
 pub trait NonPayableCallContext: MutatingCallContext {}
-
-use alloc::vec::Vec;
-use alloy_sol_types::{Panic, PanicKind, SolError};
-
-/// Represents error data when a call fails.
-#[derive(Debug, PartialEq)]
-pub enum Error {
-    /// Revert data returned by the other contract.
-    Revert(Vec<u8>),
-    /// Failure to decode the other contract's return data.
-    AbiDecodingFailed(alloy_sol_types::Error),
-}
-
-impl From<alloy_sol_types::Error> for Error {
-    fn from(err: alloy_sol_types::Error) -> Self {
-        Error::AbiDecodingFailed(err)
-    }
-}
-
-/// Encode an error.
-///
-/// This is useful so that users can use `Error` as a variant in their error
-/// types. It should not be necessary to implement this.
-pub trait MethodError {
-    /// Users should not have to call this.
-    fn encode(self) -> Vec<u8>;
-}
-
-impl MethodError for Error {
-    #[inline]
-    fn encode(self) -> Vec<u8> {
-        From::from(self)
-    }
-}
-
-impl<T: SolError> MethodError for T {
-    #[inline]
-    fn encode(self) -> Vec<u8> {
-        SolError::abi_encode(&self)
-    }
-}
-
-impl From<Error> for Vec<u8> {
-    #[allow(unused)]
-    fn from(err: Error) -> Vec<u8> {
-        match err {
-            Error::Revert(data) => data,
-            Error::AbiDecodingFailed(err) => Panic::from(PanicKind::Generic).abi_encode(),
-        }
-    }
-}
