+++ stylus-sdk/src/prelude.rs
@@ -9,7 +9,6 @@
 //! use stylus_sdk::prelude::*;
 //! ```
 
-pub use crate::host::*;
 pub use crate::storage::{Erase, SimpleStorageType, StorageType};
 pub use crate::stylus_proc::*;
 pub use crate::types::AddressVM;
